move Processing Delay test to unit tests
